Update method of recognizing samples that fail to converge
New version number: 0.1.2
This update changes the method of recognizing poorly behaved age functions (most commonly, when U, Th, or Sm is negative in a Monte Carlo simulation). Previously, the program looked to detect changes of 0.1% in the final date after 3 iterations. However, this method failed for extreme dates of ~6 Ga. This update instead recognizes inflections of the convergence direction in the Newton-Raphson method, which is a direct indication that the function is poorly behaved. These inflections are limited to 1 year to avoid problems with floating point accuracy, which can cause inflections of e-15 magnitude relative to the initial guess date.
This update also changes the way the MC samples that are poorly behaved are tracked to avoid overwriting the original list every cycle.